requirements.txt
streamlit>=1.35.0,<1.36.0
opencv-python
pandas
numpy<2
scipy
scikit-learn
lifelines
rpy2
easyocr
streamlit-drawable-canvas==0.9.3
streamlit-image-coordinates
matplotlib
Pillow
google-generativeai
